# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ensure API
  version: 2.0.0
  description: >
    Backend API for the note web app (foundation for a future deadman switch).
    Each signed-in user has at most one note of their own, stored encrypted at rest
    (feature 004-note-ownership-encryption); reads and writes are scoped to the
    authenticated user, so no endpoint can address another user's note. Gated behind
    Google SSO (feature 002-google-sso-auth): sign-in is Google-only; the access token
    is a ~1h JWT carried in an httpOnly cookie, silently refreshed while the user is
    active, with a ~24h inactivity window.
servers:
  - url: /api
    description: API root (same-origin in dev and single-instance deploy)
paths:
...
  /deadman:
    get:
      operationId: getDeadmanStatus
      summary: Get the caller's own dead-man switch status
      description: >
        Returns the authenticated caller's dead-man switch status — the current state
        (disarmed/active/grace/triggered), the configured interval and grace period, the
        absolute deadlines, a derived secondsUntilDue countdown, and the most recent
        events newest-first. A user who has never configured the switch gets a 200 with
        the default interval/grace, state "disarmed", and null deadlines. Always scoped to
        the caller — another user's switch is never returned. Requires a valid access token.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: The caller's switch status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeadmanStatus"
        "401":
          description: No valid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "UNAUTHORIZED", message: "Sign in to continue." }
  /deadman/config:
    put:
      operationId: putDeadmanConfig
      summary: Configure, arm, or disarm the caller's own dead-man switch
      description: >
        Sets the caller's check-in interval and grace period (each validated against the
        shared min/max bounds) and, via `enabled`, arms or disarms the switch. Arming
        (enabled:true) transitions a disarmed switch to "active" and sets nextCheckinDueAt
        to now + interval; disarming (enabled:false) transitions to "disarmed" and clears
        the live deadlines. An out-of-bounds interval or grace is rejected with 400
        VALIDATION_ERROR and no change. Records an armed/disarmed/config_changed event.
        Requires a valid access token.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeadmanConfigInput"
            examples:
              arm:
                value: { checkinIntervalSeconds: 604800, gracePeriodSeconds: 172800, enabled: true }
      responses:
        "200":
          description: The configuration was applied; returns the updated status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeadmanStatus"
        "400":
          description: Validation failed (interval/grace out of bounds, missing, or non-integer).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "VALIDATION_ERROR", message: "Check-in interval is out of range." }
        "401":
          description: No valid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "UNAUTHORIZED", message: "Sign in to continue." }
  /deadman/checkin:
    get:
      operationId: checkInWithToken
      summary: Check in from a passwordless email link (public, token-only)
      description: >
        PUBLIC (no session) endpoint a user opens from the one-time link embedded in a grace
        reminder email (feature 011). Authority is possession of the unguessable token only: the
        server hashes the supplied token, looks the check-in token up by hash, derives the owning
        user from that row (no caller-supplied id, no session), and — only when the token is
        valid, unused, and unexpired AND the switch is `active`/`grace` — performs the same
        check-in the dashboard does (reset last_checkin_at/next_checkin_due_at, state back to
        `active`, clear grace bookkeeping, record a `checkin` event), then marks the token used
        (single-use) and returns `checked_in`. An already-used, expired, unknown, or malformed
        token — or a token whose switch is `triggered`/`disarmed` — yields a generic
        `not_available` result disclosing nothing (the non-checkable-switch case still consumes
        the token so it cannot be replayed). The raw token is never logged or echoed.
      parameters:
        - in: query
          name: token
          required: true
          schema: { type: string }
          description: The one-time check-in token carried by the emailed link.
      responses:
        "200":
          description: The check-in outcome (checked in, or a generic not-available result).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinLinkResult"
              examples:
                checkedIn:
                  value: { status: "checked_in" }
                notAvailable:
                  value: { status: "not_available" }
    post:
      operationId: checkinDeadman
      summary: Check in ("I'm alive") on the caller's own dead-man switch
      description: >
        Resets the caller's switch on an active or grace switch — sets lastCheckinAt to
        now, moves nextCheckinDueAt to now + interval, returns the state to "active",
        clears grace bookkeeping, and records a checkin event. Rejected with 409 when the
        switch is disarmed (no live deadline) or already triggered (cannot un-fire).
        Requires a valid access token.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: The check-in was recorded; returns the updated status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeadmanStatus"
        "409":
          description: The switch is disarmed or already triggered, so a check-in does not apply.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                disarmed:
                  value: { error: "NOT_ARMED", message: "The switch is not armed." }
                triggered:
                  value: { error: "ALREADY_TRIGGERED", message: "The switch has already fired." }
        "401":
          description: No valid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "UNAUTHORIZED", message: "Sign in to continue." }
  /deadman/test-release:
    post:
      operationId: testReleaseDeadman
      summary: Preview the recipient experience by sending a test release to your own contacts
      description: >
        Mints a one-time grant to each of the caller's OWN verified contact address(es), creates
        a release tagged `manual_test`, and emails each a tokenized `${APP_BASE_URL}/r/{token}`
        link via the generic notify() dispatcher — WITHOUT changing the switch state — so a user
        can preview exactly what their contacts will receive when the switch fires (feature 010,
        the CTA feature 012 integrates). Scoped to the caller (`req.user.id`); sends only to the
        caller's own verified addresses. Returns the number of grants created. Rejected with 409
        when the caller has no verified contact (no email is sent). Requires a valid access token.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: A test release was created and emailed to the caller's own verified contact(s).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestReleaseResult"
              example: { grants: 1 }
        "401":
          description: No valid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "UNAUTHORIZED", message: "Sign in to continue." }
        "409":
          description: The caller has no verified contact, so there is nobody to send a test release to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "NO_VERIFIED_CONTACT", message: "Verify a contact before sending a test release." }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: >
        httpOnly, Secure, SameSite=Lax cookie holding the ~1h access-token JWT. Set by the Google
        callback (and by the test-login route in test mode), refreshed via POST /auth/refresh.
  schemas:
...
    CheckinLinkResult:
      type: object
      additionalProperties: false
      required: [status]
      description: >
        The outcome of opening a public passwordless check-in link (feature 011). A single
        non-disclosing shape: `checked_in` — a valid, unused, unexpired token whose switch was
        `active`/`grace` reset the clock back to `active`; `not_available` — every other case
        (missing/malformed/unknown/used/expired token, or a `triggered`/`disarmed` switch),
        disclosing nothing about whether any token/switch/user exists (fail-closed). Returned with
        HTTP 200 for BOTH outcomes.
      properties:
        status:
          type: string
          enum: [checked_in, not_available]
...
    TestReleaseResult:
      type: object
      additionalProperties: false
      required: [grants]
      description: >
        The outcome of a manual test-release (feature 010): how many one-time grants were minted
        and emailed to the caller's own verified address(es). The switch state is left unchanged.
      properties:
        grants:
          type: integer
          minimum: 0
          description: The number of grants created and emailed to the caller's own verified contacts.
    Error:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable error code.
          examples: ["UNAUTHORIZED", "VALIDATION_ERROR", "CHANNEL_NOT_SUPPORTED"]
        message:
          type: string
          description: Human-readable explanation suitable for display.
...
    DeadmanState:
      type: string
      enum: [disarmed, active, grace, triggered]
      description: >
        The current state of a user's dead-man switch. "disarmed" = off; "active" =
        armed and counting down; "grace" = a deadline was missed and the grace window is
        running (reminders sent to the user); "triggered" = the grace window also lapsed
        (in feature 008 this records the event; contact delivery is feature 010).
    DeadmanEvent:
      type: object
      additionalProperties: false
      required: [id, type, createdAt]
      description: One append-only audit entry. `detail` never contains note plaintext or tokens.
      properties:
        id:
          type: string
          description: Server-assigned unique id for this event.
        type:
          type: string
          enum:
            [
              armed,
              disarmed,
              checkin,
              entered_grace,
              reminder_sent,
              triggered,
              released,
              config_changed,
            ]
          description: >
            The kind of transition or action recorded. `released` accompanies `triggered`
            when the fired switch creates a release and emails verified contacts (feature 010);
            its detail carries only a non-sensitive grant count, never a token or plaintext.
        detail:
          type: ["string", "null"]
          description: Optional JSON detail (never note plaintext or any token value).
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp of when the event was recorded.
...
    DeadmanStatus:
      type: object
      additionalProperties: false
      required:
        [
          state,
          enabled,
          checkinIntervalSeconds,
          gracePeriodSeconds,
          lastCheckinAt,
          nextCheckinDueAt,
          graceDeadlineAt,
          secondsUntilDue,
          events,
        ]
      description: The caller's dead-man switch status, including a live countdown and recent events.
      properties:
        state: { $ref: "#/components/schemas/DeadmanState" }
        enabled:
          type: boolean
          description: Whether the switch is currently armed (enabled).
        checkinIntervalSeconds:
          type: integer
          description: The configured check-in interval, in seconds.
        gracePeriodSeconds:
          type: integer
          description: The configured grace period, in seconds.
        lastCheckinAt:
          type: ["string", "null"]
          format: date-time
          description: ISO 8601 timestamp of the last check-in/arm, or null if never.
        nextCheckinDueAt:
          type: ["string", "null"]
          format: date-time
          description: Absolute next-check-in deadline, or null when disarmed/triggered.
        graceDeadlineAt:
          type: ["string", "null"]
          format: date-time
          description: Absolute grace deadline (set in the grace state), or null otherwise.
        secondsUntilDue:
          type: ["integer", "null"]
          description: >
            Seconds until the relevant deadline (next-check-in while active, grace deadline
            while in grace), derived from the deadline and the current time; may be negative
            if a tick has not yet run, or null when there is no live deadline.
        events:
          type: array
          items:
            $ref: "#/components/schemas/DeadmanEvent"
          description: The caller's recent events, newest-first (may be empty).
    DeadmanConfigInput:
      type: object
      additionalProperties: false
      required: [checkinIntervalSeconds, gracePeriodSeconds, enabled]
      description: Configure the switch and arm/disarm it.
      properties:
        checkinIntervalSeconds:
          type: integer
          minimum: 3600
          maximum: 31536000
          description: Check-in interval in seconds (1 hour to 365 days).
        gracePeriodSeconds:
          type: integer
          minimum: 3600
          maximum: 2592000
          description: Grace period in seconds (1 hour to 30 days).
        enabled:
          type: boolean
          description: true arms the switch (→ active); false disarms it (→ disarmed).